feat(characters): add nickname and make profilePicture optional in schema

+++ src/modules/users/user.service.ts
@@ -1,10 +1,14 @@
 import { hashPassword } from "../../utils/auth/hash-password";
 import {
   ConflictError,
+  ForbiddenError,
   InternalError,
   NotFoundError,
+  UnprocessableEntityError,
 } from "../../utils/error-handler";
+import type { CharacterService } from "../characters/character.service";
 import type {
+  CreatedUser,
   CreateUserInput,
   UpdateUserInput,
   UserEntity,
@@ -13,13 +17,18 @@ import type {
 import type { UserRepository } from "./user.repository";
 
 export class UserService {
-  constructor(private readonly repo: UserRepository) {}
+  constructor(
+    private readonly repo: UserRepository,
+    private readonly characterService: CharacterService
+  ) {}
 
-  public async findMany(): Promise<UserPublic[]> {
-    return await this.repo.findAllPublic();
-  }
-
-  public async findById(userId: string): Promise<UserPublic> {
+  public async findById(
+    userId: string,
+    userToken: UserPublic
+  ): Promise<UserPublic> {
+    if (userId !== userToken.id) {
+      throw new ForbiddenError("You are not allowed to get this user");
+    }
     const user = await this.repo.findPublicById(userId);
     if (!user) {
       throw new NotFoundError("User not found", { details: { userId } });
@@ -27,78 +36,91 @@ export class UserService {
     return user;
   }
 
-  public async register(data: CreateUserInput): Promise<UserPublic> {
+  public async create(data: CreateUserInput): Promise<CreatedUser> {
     const existingUser = await this.repo.findByEmail(data.email);
     if (existingUser) {
-      throw new ConflictError("Email is already registered", {
-        details: { email: data.email },
-      });
+      throw new UnprocessableEntityError("Unable to process registration");
     }
 
-    const hashedPassword = await hashPassword(data.password);
-    const newUserEntity: UserEntity = {
+    const passwordHash = await hashPassword(data.password);
+    const newUser: UserEntity = {
       id: crypto.randomUUID(),
       name: data.name,
       email: data.email,
-      passwordHash: hashedPassword,
+      passwordHash: passwordHash,
       createdAt: new Date(),
       updatedAt: new Date(),
     };
 
-    const createdUser = await this.repo.create(newUserEntity);
-    if (!createdUser) {
+    const created = await this.repo.create(newUser);
+    if (!created) {
       throw new InternalError("Failed to create user");
     }
-    return createdUser;
+
+    const character = await this.characterService.create(created);
+
+    if (!character) {
+      throw new InternalError("Failed to character");
+    }
+
+    return { user: created, character: character };
   }
 
   public async update(
     userId: string,
-    input: UpdateUserInput
+    data: UpdateUserInput,
+    userToken: UserPublic
   ): Promise<UserPublic> {
-    const user = await this.repo.findById(userId);
-    if (!user) {
+    if (userId !== userToken.id) {
+      throw new ForbiddenError("You are not allowed to update this user");
+    }
+    const foundUser = await this.repo.findById(userId);
+    if (!foundUser) {
       throw new NotFoundError("User not found", { details: { id: userId } });
     }
 
-    if (input.email && input.email !== user.email) {
-      const existingUser = await this.repo.findByEmail(input.email);
+    if (data.email && data.email !== foundUser.email) {
+      const existingUser = await this.repo.findByEmail(data.email);
       if (existingUser && existingUser.id !== userId) {
         throw new ConflictError("Email is already registered", {
-          details: { email: input.email },
+          details: { email: data.email },
         });
       }
     }
 
-    let newPasswordHash = user.passwordHash;
-    if (input.password) {
-      newPasswordHash = await hashPassword(input.password);
-      delete input.password;
+    let newPasswordHash = foundUser.passwordHash;
+    if (data.password) {
+      newPasswordHash = await hashPassword(data.password);
+      delete data.password;
     }
 
-    const updatedEntity: UserEntity = {
-      ...user,
-      ...input,
+    const updatedUser: UserEntity = {
+      ...foundUser,
+      ...data,
       passwordHash: newPasswordHash,
-      createdAt: user.createdAt,
+      createdAt: foundUser.createdAt,
       updatedAt: new Date(),
     };
 
-    const updatedUser = await this.repo.update(userId, updatedEntity);
-    if (!updatedUser) {
+    const updated = await this.repo.update(userId, updatedUser);
+    if (!updated) {
       throw new InternalError("Failed to update user");
     }
-    return updatedUser;
+    return updated;
   }
 
-  public async delete(userId: string): Promise<void> {
+  public async delete(userId: string, userToken: UserPublic): Promise<void> {
+    if (userId !== userToken.id) {
+      throw new ForbiddenError("You are not allowed to delete this user");
+    }
+
     const existingUser = await this.repo.findById(userId);
     if (!existingUser) {
       throw new NotFoundError("User not found");
     }
 
-    const userDeleted = await this.repo.delete(userId);
-    if (!userDeleted) {
+    const deleted = await this.repo.delete(userId);
+    if (!deleted) {
       throw new NotFoundError("User not found");
     }
     return;
